+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/lib/src/events/map_event.dart
@@ -136,12 +136,40 @@ class PolylineTapEvent extends MapEvent<PolylineId> {
   PolylineTapEvent(super.mapId, super.polylineId);
 }
 
+/// An event fired when a [Polyline] path is edited by the user.
+class PolylineEditEvent extends MapEvent<PolylineId> {
+  /// Build a PolylineEdit Event triggered from the map represented by `mapId`.
+  ///
+  /// The `value` of this event is a [PolylineId] object that represents the edited Polyline.
+  /// [points] contains the updated path after the edit.
+  PolylineEditEvent(super.mapId, super.polylineId, this.points);
+
+  /// The updated list of points after the edit.
+  final List<LatLng> points;
+}
+
 /// An event fired when a [Polygon] is tapped.
 class PolygonTapEvent extends MapEvent<PolygonId> {
   /// Build an PolygonTap Event triggered from the map represented by `mapId`.
   ///
   /// The `value` of this event is a [PolygonId] object that represents the tapped Polygon.
   PolygonTapEvent(super.mapId, super.polygonId);
+}
+
+/// An event fired when a [Polygon] path is edited by the user.
+class PolygonEditEvent extends MapEvent<PolygonId> {
+  /// Build a PolygonEdit Event triggered from the map represented by `mapId`.
+  ///
+  /// The `value` of this event is a [PolygonId] object that represents the edited Polygon.
+  /// [points] contains the updated outer boundary after the edit.
+  /// [holes] contains the updated holes after the edit.
+  PolygonEditEvent(super.mapId, super.polygonId, this.points, this.holes);
+
+  /// The updated outer boundary points after the edit.
+  final List<LatLng> points;
+
+  /// The updated list of holes after the edit.
+  final List<List<LatLng>> holes;
 }
 
 /// An event fired when a [Circle] is tapped.