+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/lib/src/platform_interface/google_maps_flutter_platform.dart
@@ -327,9 +327,19 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
     throw UnimplementedError('onPolylineTap() has not been implemented.');
   }
 
+  /// A [Polyline] path has been edited by the user. Currently only supported on web.
+  Stream<PolylineEditEvent> onPolylineEdited({required int mapId}) {
+    throw UnimplementedError('onPolylineEdited() has not been implemented.');
+  }
+
   /// A [Polygon] has been tapped.
   Stream<PolygonTapEvent> onPolygonTap({required int mapId}) {
     throw UnimplementedError('onPolygonTap() has not been implemented.');
+  }
+
+  /// A [Polygon] path has been edited by the user. Currently only supported on web.
+  Stream<PolygonEditEvent> onPolygonEdited({required int mapId}) {
+    throw UnimplementedError('onPolygonEdited() has not been implemented.');
   }
 
   /// A [Circle] has been tapped.